fix: extract shared types to subtitle_types.py, fix circular import, fix watermark event deletion, add settings error logging
- Create subtitle_types.py with Change dataclass and _ms_to_str — no project imports
- Remove Change + _ms_to_str from subtitle_processor.py, import from subtitle_types instead
- Update common_fixes.py to import from subtitle_types (breaks circular dependency)
- Fix watermark stub: actually delete matched events from subs.events after recording changes
- Add logger + log warning in _build_interjection_pattern bare except
- Fix apply_mods check order: file existence before extension validation
- Use explicit format_ in pysubs2.load/save to handle empty subtitle files after all events removed
